Fixed string size limit issue
Changed to using a string list instead of a single string field as the 16 MB limitation applies to each item in the string list and not the list itself. Set the default fragment size to 16,000,000 characters.

Changed the schema to use a class so as to keep the functionality of the "content" property returning the full value. Provided a migration to handle this for existing realms.

Changed the "Realm.create" call to use the newer format as the version passing a boolean as the third property is deprecated.